+++ apps/sam-design-system-site/src/app/app.module.ts
@@ -9,24 +9,6 @@ import { FormlySampleModule } from './formly/formly.module';
 import { SdsFormlyModule } from '@sam-design-system/sam-formly'
 import { FormlyModule } from '@ngx-formly/core';
 
-
-export function minlengthValidationMessage(err, field) {
-  return `Should have atleast ${field.templateOptions.minLength} characters`;
-}
-
-export function maxlengthValidationMessage(err, field) {
-  return `This value should be less than ${field.templateOptions.maxLength} characters`;
-}
-
-export function minValidationMessage(err, field) {
-  return `This value should be more than ${field.templateOptions.min}`;
-}
-
-export function maxValidationMessage(err, field) {
-  return `This value should be less than ${field.templateOptions.max}`;
-}
-
-
 @NgModule({
   declarations: [AppComponent],
   imports: [BrowserModule,
@@ -34,15 +16,7 @@ export function maxValidationMessage(err, field) {
     RouterModule.forRoot(ROUTES),
     FormsModule,
     SdsFormlyModule,
-    FormlyModule.forRoot({
-      validationMessages: [
-        { name: 'required', message: 'This field is required' },
-        { name: 'minlength', message: minlengthValidationMessage },
-        { name: 'maxlength', message: maxlengthValidationMessage },
-        { name: 'min', message: minValidationMessage },
-        { name: 'max', message: maxValidationMessage },
-      ],
-    }),
+    FormlyModule,
     FormlySampleModule
   ],
   providers: [],
