+++ apps/sam-design-system-site/src/app/formly/formly-input/formly-input.component.html
@@ -1,15 +1,15 @@
-<!-- <form [formGroup]="form" (ngSubmit)="submit(model)">
+<form [formGroup]="form" (ngSubmit)="submit(model)">
     <formly-form [form]="form" [fields]="fields" [model]="model"></formly-form>
     <button type="submit" class="btn btn-primary submit-button" [disabled]="!form.valid">Submit</button>
-</form> -->
+</form>
 
 
 {{results | json}}
 
-<form [formGroup]="form1" (ngSubmit)="submit()">
+<!-- <form [formGroup]="form1" (ngSubmit)="submit()">
     <formly-form [model]="model1" [fields]="fields1" [options]="options1" [form]="form1"></formly-form>
     <button type="submit" class="btn btn-primary submit-button">Submit</button>
-  </form>
+  </form> -->
