+++ libs/packages/sam-formly/src/lib/formly/formly.config.ts
@@ -55,6 +55,7 @@ export const FORMLY_CONFIG: ConfigOption = {
   ],
   wrappers: [
     {name: 'form-field', component: FormlyWrapperFormFieldComponent },
+    
     {name: 'panel', component: FormlyAccordianFormFieldComponent },
   ],
 };
