+++ libs/packages/sam-formly/src/lib/formly/formly.module.ts
@@ -4,19 +4,46 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { FormlyModule } from '@ngx-formly/core';
 import { FormlySelectModule } from '@ngx-formly/core/select';
 import { FIELD_TYPE_COMPONENTS, FORMLY_CONFIG } from './formly.config';
-// import {
-//   SdsAccordionModule,
-// } from '@gsa-sam/components';
+import {
+  SdsAccordionModule,
+} from '@gsa-sam/components';
+
+export function minlengthValidationMessage(err, field) {
+  return `Should have atleast ${field.templateOptions.minLength} characters`;
+}
+
+export function maxlengthValidationMessage(err, field) {
+  return `This value should be less than ${field.templateOptions.maxLength} characters`;
+}
+
+export function minValidationMessage(err, field) {
+  return `This value should be more than ${field.templateOptions.min}`;
+}
+
+export function maxValidationMessage(err, field) {
+  return `This value should be less than ${field.templateOptions.max}`;
+}
+
+
 @NgModule({
   declarations: [
     FIELD_TYPE_COMPONENTS,
   ],
   imports: [
     CommonModule,
-    // SdsAccordionModule,
+     SdsAccordionModule,
     ReactiveFormsModule,
     FormlySelectModule,
     FormlyModule.forChild(FORMLY_CONFIG),
+    FormlyModule.forRoot({
+      validationMessages: [
+        { name: 'required', message: 'This field is required' },
+        { name: 'minlength', message: minlengthValidationMessage },
+        { name: 'maxlength', message: maxlengthValidationMessage },
+        { name: 'min', message: minValidationMessage },
+        { name: 'max', message: maxValidationMessage },
+      ],
+    })
      ],
 })
 export class SdsFormlyModule {}
