+++ libs/packages/sam-formly/src/lib/formly/types/input.ts
@@ -4,7 +4,7 @@ import { FieldType } from '@ngx-formly/core';
 @Component({
   selector: 'sam-formly-field-input',
   template: `
-  <input [class.usa-input--error]="showError" class="usa-input" id="input-success" [placeholder]="to.placeholder" name="input-success" [type]="to.inputType" [formControl]="formControl">
+  <input [class.usa-input--error]="showError" class="usa-input" [placeholder]="to.placeholder" name="input-success" [formlyAttributes]="field" [type]="to.inputType" [formControl]="formControl">
   `,
 })
 export class FormlyFieldInputComponent extends FieldType {
